#39 added auth manager.
