--- src/main/java/io/amtech/projectflow/service/token/AuthDto.java
@@ -1,11 +1,0 @@
-package io.amtech.projectflow.service.token;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-public class AuthDto {
-    private String username;
-    private String password;
-}
