--- src/main/java/io/amtech/projectflow/service/token/TokenService.java
@@ -1,14 +1,0 @@
-package io.amtech.projectflow.service.token;
-
-import org.springframework.security.core.Authentication;
-
-import javax.servlet.http.HttpServletRequest;
-
-public interface TokenService {
-    TokenDto generate(AuthDto dto);
-    TokenDto refresh(TokenRefreshDto dto);
-    void delete(String accessToken);
-    boolean isValid(String token);
-    String resolveToken(HttpServletRequest request);
-    Authentication getAuthentication(HttpServletRequest request);
-}
